#3 refactor: refactor code
